--- web/client/src/components/navigation/soro-navigation-search-content.vue
@@ -1,14 +1,0 @@
-<template>
-    <div />
-</template>
-
-<script setup lang="ts">
-</script>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-export default defineComponent({
-    name: 'SoroNavigationSearchContent',
-});
-</script>
