+++ web/client/src/components/navigation/station-search/soro-navigation-search-content.vue
@@ -1,0 +1,23 @@
+<template>
+    <div>
+        <station-search class="station-search-field" />
+    </div>
+</template>
+
+<script setup lang="ts">
+import StationSearch from './station-search.vue';
+</script>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+    name: 'SoroNavigationSearchContent',
+});
+</script>
+
+<style>
+.station-search-field {
+    padding: 3%;
+}
+</style>
